+++ HISTORY.rst
@@ -3,6 +3,10 @@
 History
 -------
 
+0.1.9 (2017-09-07)
+++++++++++++++++++
+* core.py add stat_jsonld,  add function to count triples in an JSON object
+
 0.1.8 (2017-08-16)
 ++++++++++++++++++
 * fixed summary.summarize_entity_person, handle empty nationality situation
@@ -12,7 +16,6 @@ History
 0.1.7 (2017-07-20)
 ++++++++++++++++++
 * bugfix summary.summarize_entity_person
-
 
 0.1.6 (2017-07-20)
 ++++++++++++++++++